consolidated all goal retrieval into the renamed getGoals.js file
