+++ screens/main/goal/goalsListScreen.js
@@ -7,8 +7,8 @@ import {
   TouchableOpacity,
 } from 'react-native';
 //firebase
-import auth from '@react-native-firebase/auth';
-import getGoalData from '../../../services/getData';
+// import auth from '@react-native-firebase/auth';
+import { getDisplayableGoals } from '../../../services/getGoals';
 //constants
 import { status } from '../../../services/universalConstants';
 //styles
@@ -20,26 +20,33 @@ function GoalsListScreen({ navigation }) {
   const [goalData, setGoalData] = useState([]);
 
   //retrieve goals from Firebase
-  if (auth().currentUser) {
-    let uid = auth().currentUser.uid;
-    async function getGoals() {
-      let data = await getGoalData(uid);
-      return data;
-    }
-    function setData() {
-      getGoals().then(function (goals) {
-        //checks if status is displayable
-        function checkStatus(goal) {
-          return goal.status < status.goalListDisplayable;
-        }
+  // if (auth().currentUser) {
+  //   let uid = auth().currentUser.uid;
+  //   async function getGoals() {
+  //     let data = await getGoalData(uid);
+  //     return data;
+  //   }
+  //   function setData() {
+  //     getGoals().then(function (goals) {
+  //       //checks if status is displayable
+  //       function checkStatus(goal) {
+  //         return goal.status < status.goalListDisplayable;
+  //       }
 
-        //filters array to only displayable goals
-        let displayableGoals = goals.filter(checkStatus);
-        setGoalData(displayableGoals);
-      });
-    };
-    setData();
+  //       //filters array to only displayable goals
+  //       let displayableGoals = goals.filter(checkStatus);
+  //       setGoalData(displayableGoals);
+  //     });
+  //   };
+  //   setData();
+  // }
+  
+  async function getGoals() {
+    let goals;
+    goals = await getDisplayableGoals()
+    setGoalData(goals);
   }
+  getGoals();
 
   function gotoIndividualGoal(item) {
     navigation.navigate('individualGoalDisplay', { goal: item });
