--- services/getData.js
@@ -1,33 +1,0 @@
-import firestore from '@react-native-firebase/firestore';
-
-var db = firestore().collection('Users');
-
-//retrieve data on user's goals
-async function getGoalData(uid) {
-  var goalData = [];
-  await db
-    .doc(uid)
-    .collection('goals')
-    .get()
-    .then(snap => {
-      snap.forEach(currentGoal => {
-        let docData = currentGoal.data();
-        //console.log('in getdata, chatroomid is ' + docData.chatRoomId);
-        let dataObject = {
-          title: docData.goalTitle,
-          description: docData.goalDescription,
-          milestones: docData.goalMilestones,
-          accountaBuddyId: docData.accountaBuddyId,
-          matchedGoalId: docData.matchedGoalId,
-          goalId: docData.goalId,
-          chatRoomId: docData.chatRoomId,
-          status: docData.status,
-        };
-        //add the info to goal data
-        goalData.push(dataObject);
-      });
-    });
-  return goalData;
-}
-
-export default getGoalData;
